remove unused functions, move to lib
